# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgnosTech API
  description: |
    API for AgnosTech - a tool to migrate cloud infrastructure to self-hosted Docker Compose stacks.

    ## Authentication
    Most endpoints require authentication via session cookie or Bearer token.
    Use the `/api/v1/auth/login` endpoint to obtain a session token.

    ## Error Responses
    All endpoints may return standard error responses:
    - `400 Bad Request` - Invalid request parameters
    - `401 Unauthorized` - Missing or invalid authentication
    - `404 Not Found` - Resource not found
    - `500 Internal Server Error` - Server error
    - `502 Bad Gateway` - External service connection failed
  version: 1.0.0
  contact:
    name: AgnosTech Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.homeport.local
    description: Local production server
...
paths:
...
  /api/v1/stacks/{stackId}/containers/{name}/logs:
    get:
      tags: [Containers]
      summary: Get container logs
      description: Retrieve logs from a Docker container.
      operationId: getContainerLogs
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/stackId'
        - $ref: '#/components/parameters/containerName'
        - name: tail
          in: query
          description: Number of log lines to retrieve (1-10000, default 100)
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
      responses:
        '200':
          description: Container logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: string
                    description: Log content
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT Bearer token from login
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie from login
  parameters:
    stackId:
      name: stackId
      in: path
      required: true
      description: Stack identifier (use "default" for the default stack)
      schema:
        type: string
        pattern: '^[a-zA-Z0-9][a-zA-Z0-9_.\-]*$'
        example: default
    containerName:
      name: name
      in: path
      required: true
      description: Container name
      schema:
        type: string
        maxLength: 256
        pattern: '^[a-zA-Z0-9][a-zA-Z0-9_.\-]*$'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required: [error]
...
security:
  - sessionCookie: []
  - bearerAuth: []